Regression. Fix wrong type in setting value
