+++ src/AudioSourceHost/AudioSourceHost.csproj
@@ -64,6 +64,7 @@
     <Compile Include="MarshaledEventHandler{T}.cs" />
     <Compile Include="MarshaledTaskCompletionSource.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TypeConvertHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <AdditionalFiles Include="..\stylecop.json">
