+++ src/AudioSourceHost/AudioSourceSetting.cs
@@ -61,59 +61,17 @@ namespace AudioSourceHost
         /// </summary>
         public Type SettingType { get; }
 
-        /// <summary>
-        /// Convert the value to the type of the setting.
-        /// </summary>
-        /// <param name="value">Value to convert.</param>
-        /// <returns>The value converted to the type of the setting.</returns>
-        public object ConvertToSettingType(object value)
-        {
-            if (value != null && value.GetType() != SettingType)
-            {
-                var converter = TypeDescriptor.GetConverter(value.GetType());
-                if (converter.CanConvertTo(SettingType))
-                {
-                    return converter.ConvertTo(value, SettingType);
-                }
-
-                if (TryChangeType(value, out var converted))
-                {
-                    return converted;
-                }
-
-                Logger.Error($"Unable to convert value to the setting's type. setting: `{Attribute.Name}` using value `{value}`. Setting type: `{SettingType}`, value type: `{value.GetType()}`");
-                throw new ArgumentException();
-            }
-
-            return value;
-        }
-
         private void UpdateValue(object value)
         {
             try
             {
-                var newValue = ConvertToSettingType(value);
+                var newValue = TypeConvertHelper.ConvertToType(value, SettingType);
                 _accessor[PropertyName] = newValue;
             }
             catch (Exception e)
             {
                 Logger.Error(e, $"Error occured while changing audio source settings. property: `{PropertyName}`, value: {value}");
                 throw;
-            }
-        }
-
-        private bool TryChangeType(object value, out object converted)
-        {
-            try
-            {
-                converted = Convert.ChangeType(value, SettingType);
-                return true;
-            }
-            catch (Exception e)
-            {
-                Logger.Error(e);
-                converted = null;
-                return false;
             }
         }
     }
